Move MainLayout to Structural components directory
