+++ pages/_app.js
@@ -4,11 +4,11 @@ import { PointElement, Tooltip, registry } from "chart.js"
 import dynamic from "next/dynamic"
 
 import { useCensusEmbedded, CensusEmbeddedProvider } from "@components/Contexts/CensusEmbeddedContext"
-import MainLayout from "@components/MainLayout"
 import Error_ from "@components/Message/Error_"
 import Loading from "@components/Message/Loading"
 import Sidebar from "@components/Navigation/Sidebar/Sidebar"
 import { Setup } from "@components/Setup"
+import MainLayout from "@components/Structural/Layouts/MainLayout"
 import { useBasicFetch, useFetchRuns } from "@utils/fetch"
 
 registry.add(LineElement)
